refactor: moved files to src utils and cleanup
