+++ packages/rest/samples/sampleWithApp.ts
@@ -5,7 +5,7 @@ import express from 'express'
 import { connect } from 'ngrok'
 
 import { startServer } from '../src/index'
-import { setupAgent } from '../tests/utils/agent'
+import { setupAgent } from '../src/utils/agent'
 
 const run = async () => {
   const endpoint = await connect(3001)
